requirements.txt
IMDbPY
matplotlib
pandas
